refactor(VoteView) : make VoteView folder and move VoteView related file
