+++ src/components/common/Block/BlockPopUp.tsx
@@ -3,7 +3,7 @@ import { getBlockType } from "../../../typedef/common/common.types";
 import "./css/BlockPopUp.css";
 import CommentScrollViewContainer from "./containers/CommentScrollViewContainer";
 import images from "../../../assets/images";
-import PopUpVoteViewContainer from "./containers/PopUpVoteViewContainer";
+import PopUpVoteViewContainer from "../VoteView/containers/PopUpVoteViewContainer";
 
 type Props = {
   blockDetail: getBlockType;
