+++ src/components/common/VoteView/ProgressBar.tsx
