+++ src/components/common/VoteView/VoteView.tsx
