+++ src/components/common/VoteView/css/VoteView.css
