+++ src/components/common/VoteView/css/VoteView.scss
